completed win/lose window

+++ core/src/main/java/io/github/parisajalali96/Views/ScoreBoardMenu.java
@@ -71,7 +71,7 @@ public class ScoreBoardMenu implements Screen {
         });
         survivalHeader.addListener(new ClickListener() {
             public void clicked(InputEvent event, float x, float y) {
-                users.sort(Comparator.comparingInt(User::getLongestSurvivalScore).reversed());
+                users.sort(Comparator.comparingDouble(User::getLongestSurvivalScore).reversed());
                 refreshTable();
             }
         });
